Move viewport camera to it's own file
